Rename `ProtobufWrapperTypes` to `WellKnownComparables`
