+++ mc-java-comparable/src/main/kotlin/io/spine/tools/mc/java/comparable/action/AddComparator.kt
@@ -46,7 +46,7 @@ import io.spine.protodata.render.SourceFile
 import io.spine.tools.code.Java
 import io.spine.tools.mc.java.GeneratedAnnotation
 import io.spine.tools.mc.java.comparable.ComparableMessage
-import io.spine.tools.mc.java.comparable.action.ProtobufWrapperTypes.isProtobufWrapper
+import io.spine.tools.mc.java.comparable.WellKnownComparables.isWellKnownComparable
 import io.spine.tools.mc.java.comparable.hasCompareByOption
 import io.spine.tools.psi.addFirst
 import io.spine.tools.psi.java.Environment.elementFactory
@@ -94,7 +94,7 @@ public class AddComparator(
      * 2. Enumerations (Java enums are implicitly comparable).
      * 3. Messages with [CompareByOption] option.
      * 4. External messages for which [ComparatorRegistry] has a comparator.
-     * 5. [ProtobufWrapperTypes].
+     * 5. [WellKnownComparables][io.spine.tools.mc.java.comparable.WellKnownComparables].
      *
      * @see comparingByMessage
      */
@@ -130,7 +130,8 @@ public class AddComparator(
      * having a comparator in [ComparatorRegistry].
      * 2. If the message has a [CompareByOption], then the registry should NOT have a comparator
      * for this type. Otherwise, it is unclear what to use.
-     * 3. [ProtobufWrapperTypes] are allowed to participate in comparison by default.
+     * 3. [WellKnownComparables][io.spine.tools.mc.java.comparable.WellKnownComparables]
+     * are allowed to participate in comparison by default.
      *
      * @param path The field path.
      * @param type The field type.
@@ -156,7 +157,7 @@ public class AddComparator(
             }
 
             else -> {
-                check( clazz != null && clazz.isProtobufWrapper) {
+                check( clazz != null && clazz.isWellKnownComparable) {
                     "The field `$path` has an unrecognized message type: `$type`. " +
                             "Check out the supported types in docs to `compare_by` option."
                 }
